chore: Modularise Model into PrimerFoundation
