+++ Packages/PrimerFoundation/Sources/PrimerFoundation/Model/Currency.swift
@@ -1,7 +1,7 @@
 //
 //  Currency.swift
 //
-//  Copyright © 2025 Primer API Ltd. All rights reserved. 
+//  Copyright © 2026 Primer API Ltd. All rights reserved. 
 //  Licensed under the MIT License. See LICENSE file in the project root for full license information.
 
 import Foundation
@@ -10,19 +10,24 @@ import Foundation
 public struct Currency: Codable {
     public let code: String
     public let decimalDigits: Int
-
-    enum CodingKeys: String, CodingKey {
+    
+    private enum CodingKeys: String, CodingKey {
         case code = "c"
         case decimalDigits = "m"
     }
-
-    internal var symbol: String? {
+    
+    public init(code: String, decimalDigits: Int) {
+        self.code = code
+        self.decimalDigits = decimalDigits
+    }
+    
+    public var symbol: String? {
         let localeIdentifier = Locale.identifier(fromComponents: [NSLocale.Key.currencyCode.rawValue: code])
         let locale = Locale(identifier: localeIdentifier)
         return locale.currencySymbol
     }
 
-    internal var isZeroDecimal: Bool {
+    public var isZeroDecimal: Bool {
         decimalDigits == 0
     }
 }
