+++ Packages/PrimerFoundation/Sources/PrimerFoundation/Model/PrimerCancellable.swift
@@ -1,12 +1,12 @@
 //
 //  PrimerCancellable.swift
 //
-//  Copyright © 2025 Primer API Ltd. All rights reserved. 
+//  Copyright © 2026 Primer API Ltd. All rights reserved. 
 //  Licensed under the MIT License. See LICENSE file in the project root for full license information.
 
 import Foundation
 
-protocol PrimerCancellable {
+public protocol PrimerCancellable {
     /// Cancel the associated task
     func cancel()
 }
