+++ Packages/PrimerFoundation/Sources/PrimerFoundation/Model/RetailOutletsRetail.swift
@@ -1,7 +1,7 @@
 //
 //  RetailOutletsRetail.swift
 //
-//  Copyright © 2025 Primer API Ltd. All rights reserved. 
+//  Copyright © 2026 Primer API Ltd. All rights reserved. 
 //  Licensed under the MIT License. See LICENSE file in the project root for full license information.
 
 import Foundation
