+++ Packages/PrimerFoundation/Sources/PrimerFoundation/Model/Throwable.swift
@@ -1,16 +1,16 @@
 //
 //  Throwable.swift
 //
-//  Copyright © 2025 Primer API Ltd. All rights reserved. 
+//  Copyright © 2026 Primer API Ltd. All rights reserved. 
 //  Licensed under the MIT License. See LICENSE file in the project root for full license information.
 
 import Foundation
 
-enum Throwable<T: Decodable>: Decodable {
+public enum Throwable<T: Decodable>: Decodable {
     case success(T)
     case failure(Error)
 
-    init(from decoder: Decoder) throws {
+    public init(from decoder: Decoder) throws {
         do {
             let decoded = try T(from: decoder)
             self = .success(decoded)
@@ -19,11 +19,11 @@ enum Throwable<T: Decodable>: Decodable {
         }
     }
 
-    var value: T? {
+    public var value: T? {
         switch self {
         case .failure:
             return nil
-        case .success(let value):
+        case let .success(value):
             return value
         }
     }
